Replaced `FittingDistribution` trait with the `DistributionFit` trait.
`FittingDistribution` trait is not object-safe and its `fit` method is
static. As a result of this, the kind of the constructed distribution
cannot be chosen at runtime.

It was replaced by an object-safe trait `DistributionFit`. Its method
`fit` is an instance method, which allows storing the information about
the kind of the distribution inside `self`.